backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
azure-functions==4.0.0
python-dotenv==1.0.0
requests==2.31.0
msal==1.24.0
azure-identity==1.15.0